Moved price outside CloudBalancingGenerator in order to avoid serialization of internal class by Jaxb
