+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-optaplanner/src/test/filtered-resources/kjars-sources/cloudbalance/src/main/java/org/kie/server/testing/Price.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2010 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.server.testing;
+
+public class Price {
+
+    private int hardwareValue;
+    private String description;
+    private int cost;
+
+    public Price(int hardwareValue,
+                 String description,
+                 int cost) {
+        this.hardwareValue = hardwareValue;
+        this.description = description;
+        this.cost = cost;
+    }
+
+    public int getHardwareValue() {
+        return hardwareValue;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public int getCost() {
+        return cost;
+    }
+}
